用户订单显示。

+++ src/main/java/dao/OrderDetailsDao.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface OrderDetailsDao {
     int add(OrderDetails orderDetails);
-    int delete(int g_id, int o_id);
+    int delete(int g_id, long o_id);
     int update(OrderDetails orderDetails);
     List<OrderDetails> getOrderDetailsByOId(long id);
     List<OrderDetails> findByPage(int start, int rows);
